+++ server/models/matchesModel.ts
@@ -1,0 +1,11 @@
+import mongoose from 'mongoose';
+import { Matches } from '../utils/Types';
+
+const MatchesSchema = new mongoose.Schema<Matches>({
+  userId: { type: String, required: true },
+  matchIds: { type: [String], required: true },
+});
+
+const MatchesModel = mongoose.model('Matches', MatchesSchema);
+
+export default MatchesModel;
